requirements.txt
psutil
replicate
python-dotenv
fastapi
uvicorn[standard]
